General remaster
I did a lot here. I got the website to hopefully look  a little more professional and got some JS working in the website.
This, by no means is the final design as I do need to fix the bottom the page, as well as adding more content and moving a couple things around such as the task bar. I also still need to code phone support